+++ coverage/xmlreport.py
@@ -14,12 +14,13 @@ import xml.dom.minidom
 from dataclasses import dataclass
 from typing import Any, Dict, IO, Iterable, Optional, TYPE_CHECKING, cast
 
-from coverage import __url__, __version__, files
+from coverage import __version__, files
 from coverage.misc import isolate_module, human_sorted, human_sorted_items
 from coverage.plugin import FileReporter
 from coverage.report import get_analysis_to_report
 from coverage.results import Analysis
 from coverage.types import TMorf
+from coverage.version import __url__
 
 if TYPE_CHECKING:
     from coverage import Coverage